requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arrow==1.4.0
asgiref==3.10.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.5
av==13.1.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.2
black==25.9.0
boto3==1.34.122
botocore==1.34.162
build==1.3.0
cachetools==6.2.1
canvasapi==3.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.2.1
click==8.1.8
cloup==3.0.8
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
coverage==7.9.2
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.5
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.3.9
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
docling==2.58.0
docling-core==2.49.0
docling-ibm-models==3.10.1
docling-parse==4.7.0
docx2txt==0.9
drf-spectacular==0.27.2
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
Faker==37.11.0
fastapi==0.119.1
fastapi-cli==0.0.8
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.8.0
fsspec==2025.9.0
glcontext==3.0.0
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.46.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.1.10
hiredis==3.3.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
isosurfaces==0.1.2
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kiwisolver==1.4.8
kubernetes==34.1.0
latex2mathml==3.78.1
limits==5.4.0
lxml==5.4.0
Mako==1.3.10
manim==0.19.0
manim-voiceover==0.3.7
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
mmh3==5.2.0
moderngl==5.12.0
moderngl-window==3.1.1
moto==5.1.8
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mutagen==1.47.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
ocrmac==1.0.0
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
opt_einsum==3.4.0
orjson==3.11.0
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
pgvector==0.3.6
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
polyfactory==2.22.3
posthog==5.4.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.1
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
psycopg2-binary==2.9.7
py-partiql-parser==0.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycairo==1.28.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.10.1
pydantic_core==2.41.4
pydub==0.25.1
pyglet==2.1.9
pyglm==2.8.2
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
PyLTI1p3==2.0.0
PyMuPDF==1.26.5
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Vision==12.0
pyparsing==3.2.3
pypdf==6.1.2
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.2.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==0.21.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
python-slugify==8.0.4
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.2
rapidocr==3.4.2
redis==6.4.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.0.0
rich-toolkit==0.14.8
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
s3transfer==0.10.4
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.0
screeninfo==0.8.1
semchunk==2.2.2
setuptools==80.8.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
skia-pathops==0.8.0.post2
slowapi==0.1.9
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
sox==1.5.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
srt==3.5.3
starlette==0.48.0
stevedore==5.4.1
stripe==13.0.1
structlog==24.1.0
svgelements==1.9.6
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.12
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.21.0
virtualenv==20.31.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlsxwriter==3.2.9
xmltodict==0.14.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
